requirements.txt
streamlit>=1.32.0
gpt4all>=2.8.2
python-dotenv>=1.0.1
SpeechRecognition>=3.10.1
sounddevice>=0.5.2
numpy>=1.26.4
fastapi>=0.109.1
uvicorn>=0.27.0
yfinance>=0.2.36
pandas>=2.2.0
faiss-cpu>=1.7.4
beautifulsoup4>=4.12.3
requests>=2.31.0
crewai>=0.5.0
python-multipart>=0.0.9
openai-whisper
pyttsx3>=2.90
aiohttp>=3.9.3
httpx>=0.26.0
pytz